Core Team: suggested changes
Add a primary link. 

Remove the 'Items' heading, promote the subsequent headings. 

Core Team, core team … consistency with capitalisation. I suggest 'the Core Team', 'the Team', or 'Core' (as in the bullet points). 

Uppercase T for The FreeBSD Foundation and uppercase B for its Board. 

Chronological order for dated things in the paragraph about the Charter. 

Add a URL for Matrix. 

For each bullet point, below a heading, that is a complete sentence: end with a full stop. Consistent with changes to another report for this quarter.